refactor: AdmxCache search facade extraction
