+++ PolicyPlusCore/Core/AdmxCache.cs
@@ -1533,993 +1533,14 @@ VALUES(@pid,@culture,@dname,@desc,@cat,@kana,@pres)";
     ) =>
         await SearchAsync(query, cultures, fields, andMode: false, limit, ct).ConfigureAwait(false);
 
-    public async Task<IReadOnlyList<PolicyHit>> SearchAsync(
+    public Task<IReadOnlyList<PolicyHit>> SearchAsync(
         string query,
         IReadOnlyList<string> cultures,
         SearchFields fields,
         bool andMode,
         int limit = 50,
         CancellationToken ct = default
-    )
-    {
-        if (cultures == null || cultures.Count == 0)
-            cultures = new[] { CultureInfo.CurrentUICulture.Name };
-        if (string.IsNullOrWhiteSpace(query) || fields == SearchFields.None)
-            return Array.Empty<PolicyHit>();
-
-        var qExact = query;
-        var qPrefix = query + "%";
-        var qStrict = TextNormalization.NormalizeStrict(query);
-        var norm = TextNormalization.ToNGramTokens(qStrict);
-        var loose = TextNormalization.ToNGramTokens(
-            TextNormalization.NormalizeLooseFromStrict(qStrict)
-        );
-
-        bool useId = (fields & SearchFields.Id) != 0;
-        bool useReg = (fields & SearchFields.Registry) != 0 && LooksLikeRegistryQuery(query);
-        bool useName = (fields & SearchFields.Name) != 0;
-        bool useDesc = (fields & SearchFields.Description) != 0;
-        bool enableFts = useName || useDesc || useReg || useId;
-        bool shortQuery = qStrict.Length <= 2; // Special-case very short queries (<=2) because they create disproportionate noise.
-        if (shortQuery)
-        {
-            // Suppress N‑gram FTS for very short queries; restrict to Exact / Prefix / RegistryExact to prevent explosive candidate growth.
-            // Aligns with non-cache scoring where 2-char mid-word matches score only 20 and rarely surface near the top.
-            enableFts = false;
-        }
-
-        var ftsNormCols = new List<string>();
-        if (useName)
-            ftsNormCols.Add("title_norm");
-        if (useDesc)
-            ftsNormCols.Add("desc_norm");
-        if (useId && LooksLikeIdQuery(query))
-            ftsNormCols.Add("tags");
-        if (useReg)
-            ftsNormCols.Add("registry_path");
-        var ftsLooseCols = new List<string>();
-        if (useName)
-            ftsLooseCols.Add("title_loose");
-        if (useDesc)
-            ftsLooseCols.Add("desc_loose");
-        if (useId && LooksLikeIdQuery(query))
-            ftsLooseCols.Add("tags");
-        if (useReg)
-            ftsLooseCols.Add("registry_path");
-
-        static string EscapeSqlSingle(string s) => s.Replace("'", "''");
-        static IEnumerable<string> SplitTokens(string grams) =>
-            grams.Split(
-                ' ',
-                StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries
-            );
-        static string SanitizeTok(string t)
-        {
-            if (string.IsNullOrEmpty(t))
-                return string.Empty;
-            var sb = new StringBuilder(t.Length);
-            foreach (var ch in t)
-                if (char.IsLetterOrDigit(ch))
-                    sb.Append(ch);
-            return sb.ToString();
-        }
-        static string BuildFtsOr(List<string> cols, IEnumerable<string> rawTokens)
-        {
-            var toks = rawTokens
-                .Select(SanitizeTok)
-                .Where(x => !string.IsNullOrWhiteSpace(x))
-                .ToList();
-            if (cols.Count == 0 || toks.Count == 0)
-                return string.Empty;
-            // For each token build (col1:(tok) OR col2:(tok) ...) and join all with OR
-            var perTok = new List<string>(toks.Count);
-            foreach (var t in toks)
-                perTok.Add("(" + string.Join(" OR ", cols.Select(c => c + ":(" + t + ")")) + ")");
-            return string.Join(" OR ", perTok);
-        }
-        static string BuildFtsAnd(List<string> cols, IEnumerable<string> rawTokens)
-        {
-            var toks = rawTokens
-                .Select(SanitizeTok)
-                .Where(x => !string.IsNullOrWhiteSpace(x))
-                .ToList();
-            if (cols.Count == 0 || toks.Count == 0)
-                return string.Empty;
-            var per = new List<string>(toks.Count);
-            foreach (var t in toks)
-                per.Add("(" + string.Join(" OR ", cols.Select(c => c + ":(" + t + ")")) + ")");
-            return string.Join(" AND ", per);
-        }
-        var normTokens = SplitTokens(norm);
-        var looseTokens = SplitTokens(loose);
-
-        // Build expression requiring ALL tokens for a segment: ( (c1:(g1) OR c2:(g1)) AND (c1:(g2) OR c2:(g2)) ... )
-        static string BuildPerSegmentAllGrams(
-            List<string> cols,
-            IEnumerable<string> rawTokens,
-            int maxGrams = 8
-        )
-        {
-            if (cols.Count == 0)
-                return string.Empty;
-            var toks = rawTokens
-                .Select(SanitizeTok)
-                .Where(x => !string.IsNullOrWhiteSpace(x))
-                .Distinct()
-                .Take(maxGrams) // cap to avoid very large FTS expressions on long segments
-                .ToList();
-            if (toks.Count == 0)
-                return string.Empty;
-            if (toks.Count == 1)
-                return "(" + string.Join(" OR ", cols.Select(c => c + ":(" + toks[0] + ")")) + ")";
-            var per = new List<string>(toks.Count);
-            foreach (var t in toks)
-                per.Add("(" + string.Join(" OR ", cols.Select(c => c + ":(" + t + ")")) + ")");
-            return "(" + string.Join(" AND ", per) + ")";
-        }
-        static string BuildFtsSegmentedAnd(List<string> cols, IEnumerable<string> segments)
-        {
-            if (cols.Count == 0)
-                return string.Empty;
-            var segExprs = new List<string>();
-            foreach (var seg in segments)
-            {
-                if (string.IsNullOrWhiteSpace(seg))
-                    continue;
-                var grams = TextNormalization.ToNGramTokens(seg);
-                if (string.IsNullOrWhiteSpace(grams))
-                {
-                    // Fallback: if the segment is 1 char (cannot produce 2-gram), use the raw char directly.
-                    if (seg.Length == 1)
-                    {
-                        var tok = SanitizeTok(seg);
-                        if (!string.IsNullOrWhiteSpace(tok))
-                            grams = tok;
-                    }
-                }
-                var toks = string.IsNullOrEmpty(grams)
-                    ? Array.Empty<string>()
-                    : grams.Split(
-                        ' ',
-                        StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries
-                    );
-                if (toks.Length > 0)
-                {
-                    // Heuristic: if any 3-grams exist, drop 2-grams to improve precision (common 2-grams cause broad matches in Japanese).
-                    bool hasTri = toks.Any(t => t.Length == 3);
-                    if (hasTri)
-                    {
-                        var filtered = toks.Where(t => t.Length == 3).Take(10).ToArray(); // cap per segment
-                        if (filtered.Length > 0)
-                            toks = filtered;
-                    }
-                }
-                if (toks.Length == 0)
-                    continue;
-                var expr = BuildPerSegmentAllGrams(cols, toks);
-                if (!string.IsNullOrWhiteSpace(expr))
-                    segExprs.Add(expr); // already wrapped
-            }
-            if (segExprs.Count == 0)
-                return string.Empty;
-            return string.Join(" AND ", segExprs);
-        }
-
-        string matchNorm;
-        string matchLoose;
-        if (andMode && qStrict.IndexOf(' ') >= 0)
-        {
-            var segmentsStrict = qStrict.Split(
-                ' ',
-                StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries
-            );
-            matchNorm = BuildFtsSegmentedAnd(ftsNormCols, segmentsStrict);
-            // For loose, re-normalize each strict segment into loose form then build.
-            var segmentsLoose = segmentsStrict.Select(s =>
-                TextNormalization.NormalizeLooseFromStrict(s)
-            );
-            matchLoose = BuildFtsSegmentedAnd(ftsLooseCols, segmentsLoose);
-            // If segmentation produced nothing (e.g., all 1-char segments), fall back to original behavior.
-            if (string.IsNullOrWhiteSpace(matchNorm))
-                matchNorm = BuildFtsAnd(ftsNormCols, normTokens);
-            if (string.IsNullOrWhiteSpace(matchLoose))
-                matchLoose = BuildFtsAnd(ftsLooseCols, looseTokens);
-        }
-        else
-        {
-            // Precision tweak: if searching only by Name (no Id/Registry/Description) and not explicitly in AND mode,
-            // require all grams (AND) instead of any (OR). The previous OR behavior caused false positives where a token
-            // chosen from description (not present in display name) shared a single common 2-gram with the title, making
-            // the Name-only search incorrectly match. This aligns with integration test expectation that a description-only
-            // token will not hit a Name-only search.
-            bool nameOnly = useName && !useDesc && !useReg && !useId;
-            // Helper to detect CJK-ish chars (broad ranges: CJK Unified, Hiragana, Katakana, compatibility).
-            static bool IsCjk(char ch) =>
-                (ch >= '\u3040' && ch <= '\u30FF') // Hiragana + Katakana
-                || (ch >= '\u3400' && ch <= '\u9FFF') // CJK Unified Ext A + Basic
-                || (ch >= '\uF900' && ch <= '\uFAFF'); // CJK Compatibility Ideographs
-            bool isSingle = qStrict.IndexOf(' ') < 0;
-            bool isAsciiWord =
-                isSingle
-                && qStrict.Length >= 8
-                && qStrict.All(ch => ch < 128 && (char.IsLetterOrDigit(ch)));
-            // For CJK, even short words can create noisy partial matches; treat a single token length >=3 as a precise-search case.
-            bool isCjkWord = isSingle && qStrict.Length >= 3 && qStrict.Any(IsCjk);
-            bool forcePreciseSingleToken = !andMode && (isAsciiWord || isCjkWord);
-            if (shortQuery)
-            {
-                // Dummy value because FTS is already disabled for short queries.
-                matchNorm = string.Empty;
-                matchLoose = string.Empty;
-            }
-            else if (!andMode && nameOnly)
-            {
-                matchNorm = BuildFtsAnd(ftsNormCols, normTokens);
-                matchLoose = BuildFtsAnd(ftsLooseCols, looseTokens);
-            }
-            else if (forcePreciseSingleToken)
-            {
-                matchNorm = BuildFtsAnd(ftsNormCols, normTokens);
-                matchLoose = BuildFtsAnd(ftsLooseCols, looseTokens);
-            }
-            else
-            {
-                matchNorm = andMode
-                    ? BuildFtsAnd(ftsNormCols, normTokens)
-                    : BuildFtsOr(ftsNormCols, normTokens);
-                matchLoose = andMode
-                    ? BuildFtsAnd(ftsLooseCols, looseTokens)
-                    : BuildFtsOr(ftsLooseCols, looseTokens);
-            }
-        }
-        var matchNormEsc = EscapeSqlSingle(matchNorm);
-        var matchLooseEsc = EscapeSqlSingle(matchLoose);
-
-        // Precompute query gram set (strict) for overlap-based secondary weighting (tie-breaking Name vs Description dominance).
-        var queryGramSet = new HashSet<string>(normTokens, StringComparer.Ordinal);
-
-        using var conn = _store.OpenConnection();
-        await conn.OpenAsync(ct).ConfigureAwait(false);
-        var sb = new StringBuilder();
-        bool phraseMode = !andMode && qStrict.Contains(' ');
-        if (phraseMode)
-        {
-            // Multi-word (non-AND) queries: limit to phrase substring matches; disable broad n-gram FTS.
-            enableFts = false;
-        }
-        // phraseMode: Use ONLY LIKE-based phrase substring matching. Previous behavior considered combining with FTS, but it is now fully disabled for noise suppression.
-        string? phraseLike = null;
-        if (phraseMode)
-        {
-            // qStrict is already lower-cased and spaces collapsed.
-            // Escape SQL wildcard characters to avoid accidental broadening.
-            var esc = qStrict.Replace("%", "\\%").Replace("_", "\\_");
-            phraseLike = "%" + esc + "%"; // substring search
-        }
-        // Build prioritized culture list (order preserved from caller)
-        sb.AppendLine("WITH CulturePref AS (");
-        for (int i = 0; i < cultures.Count; i++)
-        {
-            if (i > 0)
-                sb.AppendLine("UNION ALL");
-            sb.Append("SELECT ")
-                .Append(i.ToString(CultureInfo.InvariantCulture))
-                .Append(" AS prio, @c")
-                .Append(i)
-                .Append(" AS culture")
-                .AppendLine();
-        }
-        sb.AppendLine(") , ExactPrefix AS (");
-        sb.AppendLine(
-            "  SELECT NULL AS id, NULL AS culture, '' AS display_name, -1 AS score WHERE 1=0"
-        );
-        if (useId)
-        {
-            sb.AppendLine("  UNION ALL");
-            sb.AppendLine(
-                "  SELECT p.id, s.culture, s.display_name, 1200 FROM Policies p JOIN PolicyI18n s ON s.policy_id=p.id WHERE p.policy_name=@q_exact AND s.culture IN (SELECT culture FROM CulturePref)"
-            );
-            sb.AppendLine("  UNION ALL");
-            sb.AppendLine(
-                "  SELECT p.id, s.culture, s.display_name, 300 FROM Policies p JOIN PolicyI18n s ON s.policy_id=p.id WHERE p.policy_name LIKE @q_prefix AND s.culture IN (SELECT culture FROM CulturePref)"
-            );
-        }
-        if (useReg)
-        {
-            sb.AppendLine("  UNION ALL");
-            sb.AppendLine(
-                "  SELECT p.id, s.culture, s.display_name, 1000 FROM Policies p JOIN PolicyI18n s ON s.policy_id=p.id WHERE (p.hive||'\\\\'||p.reg_key||'\\\\'||p.reg_value)=@q_exact AND s.culture IN (SELECT culture FROM CulturePref)"
-            );
-        }
-        sb.AppendLine(") , Phrase AS (");
-        if (phraseMode)
-        {
-            sb.AppendLine(
-                "  SELECT p.id, s.culture, s.display_name, 160 AS score FROM Policies p JOIN PolicyI18n s ON s.policy_id=p.id WHERE s.culture IN (SELECT culture FROM CulturePref) AND ("
-            );
-            // Lower-case fields for comparison (display_name, explain_text); qStrict already lower.
-            sb.AppendLine(
-                "      LOWER(s.display_name) LIKE @phraseLike ESCAPE '\\' OR LOWER(s.explain_text) LIKE @phraseLike ESCAPE '\\'"
-            );
-            sb.AppendLine("  )");
-        }
-        else
-        {
-            sb.AppendLine(
-                "  SELECT NULL AS id, NULL AS culture, '' AS display_name, -1 AS score WHERE 1=0"
-            );
-        }
-        sb.AppendLine(") , F1 AS (");
-        sb.AppendLine(
-            "  SELECT m.policy_id AS id, m.culture, s.display_name, 100 AS score FROM PolicyIndex JOIN PolicyIndexMap m ON m.rowid=PolicyIndex.rowid JOIN PolicyI18n s ON s.policy_id=m.policy_id AND s.culture=m.culture WHERE @enableFts=1 AND m.culture IN (SELECT culture FROM CulturePref)"
-        );
-        if (enableFts && ftsNormCols.Count > 0 && !string.IsNullOrWhiteSpace(matchNorm))
-            sb.Append("    AND PolicyIndex MATCH '").Append(matchNormEsc).AppendLine("'");
-        else
-            sb.AppendLine("    AND (0)");
-        sb.AppendLine(") , F2 AS (");
-        sb.AppendLine(
-            "  SELECT m.policy_id AS id, m.culture, s.display_name, 60 AS score FROM PolicyIndex JOIN PolicyIndexMap m ON m.rowid=PolicyIndex.rowid JOIN PolicyI18n s ON s.policy_id=m.policy_id AND s.culture=m.culture WHERE @enableFts=1 AND m.culture IN (SELECT culture FROM CulturePref)"
-        );
-        if (enableFts && ftsLooseCols.Count > 0 && !string.IsNullOrWhiteSpace(matchLoose))
-            sb.Append("    AND PolicyIndex MATCH '").Append(matchLooseEsc).AppendLine("'");
-        else
-            sb.AppendLine("    AND (0)");
-        sb.AppendLine(") , Candidates AS (");
-        sb.AppendLine("  SELECT id,culture,display_name,score FROM ExactPrefix");
-        sb.AppendLine("  UNION ALL SELECT id,culture,display_name,score FROM Phrase");
-        sb.AppendLine("  UNION ALL SELECT id,culture,display_name,score FROM F1");
-        sb.AppendLine("  UNION ALL SELECT id,culture,display_name,score FROM F2");
-        // Filtered: allow primary always; allow second always (if provided); allow fallback cultures only when primary row absent for that policy.
-        sb.AppendLine(") , Filtered AS (");
-        sb.AppendLine("  SELECT c.* FROM Candidates c");
-        sb.AppendLine("  WHERE c.culture = @primaryCulture");
-        sb.AppendLine("     OR (@secondCulture IS NOT NULL AND c.culture = @secondCulture)");
-        sb.AppendLine("     OR ( c.culture <> @primaryCulture");
-        sb.AppendLine("          AND (@secondCulture IS NULL OR c.culture <> @secondCulture)");
-        sb.AppendLine(
-            "          AND NOT EXISTS (SELECT 1 FROM PolicyI18n px WHERE px.policy_id = c.id AND px.culture = @primaryCulture)"
-        );
-        sb.AppendLine(
-            "          AND ( @secondCulture IS NULL OR NOT EXISTS (SELECT 1 FROM PolicyI18n sx WHERE sx.policy_id = c.id AND sx.culture = @secondCulture) )"
-        );
-        sb.AppendLine("        )");
-        sb.AppendLine(") , Ranked AS (");
-        sb.AppendLine(
-            "  SELECT c.id,c.culture,c.display_name,c.score,cp.prio, ROW_NUMBER() OVER (PARTITION BY c.id ORDER BY cp.prio ASC) AS rnk FROM Filtered c JOIN CulturePref cp ON cp.culture=c.culture"
-        );
-        sb.AppendLine(
-            ") SELECT R.id,R.culture, (SELECT ns||':'||policy_name FROM Policies P WHERE P.id=R.id) AS unique_id,"
-        );
-        sb.AppendLine("       R.score, R.display_name,");
-        sb.AppendLine(
-            "       (SELECT hive||'\\\\'||reg_key||'\\\\'||reg_value FROM Policies P2 WHERE P2.id=R.id) AS registry_path,"
-        );
-        sb.AppendLine(
-            "       (SELECT product_hint FROM Policies P3 WHERE P3.id=R.id) AS product_hint,"
-        );
-        sb.AppendLine(
-            "       (SELECT value_type FROM Policies P4 WHERE P4.id=R.id) AS value_type,"
-        );
-        sb.AppendLine(
-            "       (SELECT explain_text FROM PolicyI18n PX WHERE PX.policy_id=R.id AND PX.culture=R.culture) AS explain_text"
-        );
-        sb.AppendLine("  FROM Ranked R WHERE R.rnk = 1 ORDER BY R.score DESC LIMIT @limit; ");
-
-        // Expand internal fetch limit when description is included so that post-reordering can surface Name/ID hits
-        // even if raw SQL top-N was dominated by description-only matches.
-        int internalLimit = limit;
-        if (useDesc && (useName || useId || useReg))
-        {
-            try
-            {
-                checked
-                {
-                    internalLimit = Math.Min(limit * 3, 300);
-                }
-            }
-            catch
-            {
-                internalLimit = Math.Min(limit * 2, 300);
-            }
-        }
-        // Collect hits with phrase flag (for phase ordering when phraseMode=true)
-        var hits = new List<(PolicyHit Hit, bool Phrase)>(Math.Min(internalLimit, 512));
-        using var cmd = conn.CreateCommand();
-        cmd.CommandText = sb.ToString();
-        cmd.Parameters.AddWithValue("@q_exact", qExact);
-        cmd.Parameters.AddWithValue("@q_prefix", qPrefix);
-        cmd.Parameters.AddWithValue("@limit", internalLimit);
-        cmd.Parameters.AddWithValue("@enableFts", enableFts ? 1 : 0);
-        if (phraseMode && phraseLike != null)
-        {
-            var pPhrase = cmd.CreateParameter();
-            pPhrase.ParameterName = "@phraseLike";
-            pPhrase.Value = phraseLike;
-            cmd.Parameters.Add(pPhrase);
-        }
-        for (int i = 0; i < cultures.Count; i++)
-            cmd.Parameters.AddWithValue("@c" + i, NormalizeCultureName(cultures[i]));
-        // Primary culture = first in list; second = second element if exists; others treated as fallback candidates.
-        var primaryCulture =
-            cultures.Count > 0
-                ? NormalizeCultureName(cultures[0])
-                : CultureInfo.CurrentUICulture.Name;
-        string? secondCulture = cultures.Count > 1 ? NormalizeCultureName(cultures[1]) : null;
-        if (
-            secondCulture != null
-            && string.Equals(secondCulture, primaryCulture, StringComparison.OrdinalIgnoreCase)
-        )
-        {
-            // Explicit duplicate primary acts as placeholder second – treat as absent so fallback suppression logic engages correctly.
-            secondCulture = null;
-        }
-        cmd.Parameters.AddWithValue("@primaryCulture", primaryCulture);
-        var pSecond = cmd.CreateParameter();
-        pSecond.ParameterName = "@secondCulture";
-        if (secondCulture == null)
-            pSecond.Value = DBNull.Value;
-        else
-            pSecond.Value = secondCulture;
-        cmd.Parameters.Add(pSecond);
-        using var rdr = await cmd.ExecuteReaderAsync(ct).ConfigureAwait(false);
-        while (await rdr.ReadAsync(ct).ConfigureAwait(false))
-        {
-            var pid = rdr.GetInt64(0);
-            var cul = rdr.GetString(1);
-            var uid = rdr.IsDBNull(2) ? string.Empty : rdr.GetString(2);
-            var score = rdr.GetDouble(3);
-            var dname = rdr.IsDBNull(4) ? string.Empty : rdr.GetString(4);
-            var reg = rdr.IsDBNull(5) ? string.Empty : rdr.GetString(5);
-            var prod = rdr.IsDBNull(6) ? string.Empty : rdr.GetString(6);
-            var vtype = rdr.IsDBNull(7) ? string.Empty : rdr.GetString(7);
-            var explainText = rdr.IsDBNull(8) ? string.Empty : rdr.GetString(8);
-            // Precision tightening: for long single-token queries we require a full substring hit in at least one enabled field
-            // (ID / Name / Registry path / Description). This suppresses cases where high-frequency 2-grams inside a long token
-            // all appear scattered across a long description, yielding false positives.
-            bool singleToken = qStrict.IndexOf(' ') < 0;
-            bool singleAsciiLong =
-                singleToken
-                && qStrict.Length >= 3 // lowered from 8 -> 3 to suppress medium-length noisy n-gram hits (e.g., 'Cache')
-                && qStrict.All(ch => ch < 128 && (char.IsLetterOrDigit(ch)));
-            bool singleCjkLong =
-                singleToken
-                && qStrict.Length >= 3
-                && qStrict.Any(ch =>
-                    (ch >= '\u3040' && ch <= '\u30FF')
-                    || (ch >= '\u3400' && ch <= '\u9FFF')
-                    || (ch >= '\uF900' && ch <= '\uFAFF')
-                );
-            if ((singleAsciiLong || singleCjkLong))
-            {
-                var qLowerSub = qStrict;
-                bool hitId = false,
-                    hitName = false,
-                    hitReg = false,
-                    hitDesc = false;
-                if (
-                    useId
-                    && !string.IsNullOrEmpty(uid)
-                    && uid.IndexOf(qLowerSub, StringComparison.OrdinalIgnoreCase) >= 0
-                )
-                    hitId = true;
-                if (
-                    useName
-                    && !string.IsNullOrEmpty(dname)
-                    && dname.IndexOf(qLowerSub, StringComparison.OrdinalIgnoreCase) >= 0
-                )
-                    hitName = true;
-                if (
-                    useReg
-                    && !string.IsNullOrEmpty(reg)
-                    && reg.IndexOf(qLowerSub, StringComparison.OrdinalIgnoreCase) >= 0
-                )
-                    hitReg = true;
-                if (
-                    useDesc
-                    && !string.IsNullOrEmpty(explainText)
-                    && explainText.IndexOf(qLowerSub, StringComparison.OrdinalIgnoreCase) >= 0
-                )
-                    hitDesc = true;
-
-                if (!(hitId || hitName || hitReg || hitDesc))
-                {
-                    // No full substring across any enabled field -> discard.
-                    continue;
-                }
-
-                // Additional precision rule: if the substring occurs only in the Description (explanation text)
-                // and not in ID / DisplayName / Registry path, drop it for single-token queries.
-                // Rationale: description-only matches massively inflate result counts (e.g., 'Cache') while
-                // non-cache search heuristics (ScoreMatch) effectively prioritize surface fields, keeping
-                // recall manageable. This narrows ~1000 widespread description hits down toward the
-                // tighter set (~100) that also surface the term in a primary identifying field.
-                if (hitDesc && !(hitId || hitName || hitReg))
-                {
-                    // Allow short (<=3) ASCII tokens to still use description for recall – here we already
-                    // know length >=3 (ASCII) or >=3 (CJK). We enforce for ASCII length >=4 and for CJK length >=3.
-                    bool ascii = qStrict.All(ch => ch < 128 && (char.IsLetterOrDigit(ch)));
-                    if ((ascii && qStrict.Length >= 4) || (!ascii && qStrict.Length >= 3))
-                        continue; // description-only -> discard
-                }
-            }
-            // Boost near-exact large overlap: if user entered a long single token and it matches start of ID or DisplayName.
-            // Boost for precise single-token searches (>=8 chars ASCII) so exact/substring matches outrank loose n-gram overlaps.
-            bool boostAscii =
-                qStrict.IndexOf(' ') < 0
-                && qStrict.Length >= 8
-                && qStrict.All(ch => ch < 128 && (char.IsLetterOrDigit(ch)));
-            bool boostCjk =
-                qStrict.IndexOf(' ') < 0
-                && qStrict.Length >= 3
-                && qStrict.Any(ch =>
-                    (ch >= '\u3040' && ch <= '\u30FF')
-                    || (ch >= '\u3400' && ch <= '\u9FFF')
-                    || (ch >= '\uF900' && ch <= '\uFAFF')
-                );
-            if (boostAscii || boostCjk)
-            {
-                var qLower = qStrict.ToLowerInvariant();
-                if (!string.IsNullOrEmpty(uid) && uid.ToLowerInvariant().Contains(qLower))
-                    score += 240; // elevate ID-contained hits above loose grams
-                else if (!string.IsNullOrEmpty(dname) && dname.ToLowerInvariant().Contains(qLower))
-                    score += 140; // display name match secondary boost
-            }
-            // Field priority weighting: ID/Name > registry value > registry key > description (only within normal FTS/Phrase score band)
-            if (score >= 20 && score <= 400)
-            {
-                var qLower2 = qStrict;
-                bool idHit =
-                    !string.IsNullOrEmpty(uid)
-                    && uid.IndexOf(qLower2, StringComparison.OrdinalIgnoreCase) >= 0;
-                bool nameHit =
-                    !string.IsNullOrEmpty(dname)
-                    && dname.IndexOf(qLower2, StringComparison.OrdinalIgnoreCase) >= 0;
-                bool regHit =
-                    !string.IsNullOrEmpty(reg)
-                    && reg.IndexOf(qLower2, StringComparison.OrdinalIgnoreCase) >= 0;
-                bool descHit =
-                    !string.IsNullOrEmpty(explainText)
-                    && explainText.IndexOf(qLower2, StringComparison.OrdinalIgnoreCase) >= 0;
-                bool regValueHit = false;
-                if (regHit)
-                {
-                    try
-                    {
-                        var parts = reg.Split('\\');
-                        if (parts.Length > 0)
-                        {
-                            var last = parts[^1];
-                            if (
-                                !string.IsNullOrEmpty(last)
-                                && last.IndexOf(qLower2, StringComparison.OrdinalIgnoreCase) >= 0
-                            )
-                                regValueHit = true;
-                        }
-                    }
-                    catch { }
-                }
-                int add = 0;
-                if (idHit || nameHit)
-                    add = 55; // top tier
-                else if (regValueHit)
-                    add = 40; // mid tier
-                else if (regHit)
-                    add = 28; // key path only
-                else if (descHit)
-                    add = 12; // lowest tier
-                score += add;
-
-                // Secondary weighting: relative overlap of n-grams with the query (Name prioritized). Helps when there is no exact substring match (e.g., minor typos).
-                // Compute only for small result set; limit already caps to 'limit'.
-                try
-                {
-                    static bool IsAsciiLetterOrDigit(char c) =>
-                        (c >= '0' && c <= '9') || (c >= 'a' && c <= 'z') || (c >= 'A' && c <= 'Z');
-                    static bool WordBoundaryContains(string src, string needle)
-                    {
-                        if (string.IsNullOrEmpty(src) || string.IsNullOrEmpty(needle))
-                            return false;
-                        var lowerSrc = src.ToLowerInvariant();
-                        var lowerNeedle = needle.ToLowerInvariant();
-                        int idx = 0;
-                        while (
-                            (idx = lowerSrc.IndexOf(lowerNeedle, idx, StringComparison.Ordinal))
-                            >= 0
-                        )
-                        {
-                            bool startOk = idx == 0 || !IsAsciiLetterOrDigit(lowerSrc[idx - 1]);
-                            int endPos = idx + lowerNeedle.Length;
-                            bool endOk =
-                                endPos >= lowerSrc.Length
-                                || !IsAsciiLetterOrDigit(lowerSrc[endPos]);
-                            if (startOk && endOk)
-                                return true;
-                            idx = idx + 1;
-                        }
-                        return false;
-                    }
-                    int nameOverlap = 0;
-                    int descOverlap = 0;
-                    if (!string.IsNullOrEmpty(dname))
-                    {
-                        var dnStrict = TextNormalization.NormalizeStrict(dname);
-                        var dnTok = TextNormalization.ToNGramTokens(dnStrict);
-                        if (!string.IsNullOrWhiteSpace(dnTok))
-                        {
-                            foreach (
-                                var t in dnTok.Split(
-                                    ' ',
-                                    StringSplitOptions.RemoveEmptyEntries
-                                        | StringSplitOptions.TrimEntries
-                                )
-                            )
-                                if (t.Length > 0 && queryGramSet.Contains(t))
-                                    nameOverlap++;
-                        }
-                    }
-                    if (!string.IsNullOrEmpty(explainText))
-                    {
-                        var exStrict = TextNormalization.NormalizeStrict(explainText);
-                        var exTok = TextNormalization.ToNGramTokens(exStrict);
-                        if (!string.IsNullOrWhiteSpace(exTok))
-                        {
-                            foreach (
-                                var t in exTok.Split(
-                                    ' ',
-                                    StringSplitOptions.RemoveEmptyEntries
-                                        | StringSplitOptions.TrimEntries
-                                )
-                            )
-                                if (t.Length > 0 && queryGramSet.Contains(t))
-                                    descOverlap++;
-                        }
-                    }
-                    if (nameOverlap > 0 || descOverlap > 0)
-                    {
-                        // Weight: each name overlap stronger than description; cap to avoid runaway.
-                        if (nameOverlap > 0)
-                            score += Math.Min(60, nameOverlap * 6); // up to +60 (unchanged)
-                        if (descOverlap > 0)
-                            score += Math.Min(20, descOverlap * 2); // cap lower (max +20) to reduce description dominance
-                        // Word-boundary exact match in Name or ID gets an extra bump (strong intent signal)
-                        if (
-                            WordBoundaryContains(dname ?? string.Empty, qLower2)
-                            || WordBoundaryContains(uid ?? string.Empty, qLower2)
-                        )
-                            score += 35; // boundary exact bonus
-                        // Description-only penalty stronger now
-                        if (!idHit && !nameHit && nameOverlap == 0 && descOverlap > 0 && !regHit)
-                            score -= 35; // stronger penalty so description-only rarely tops list
-                    }
-                }
-                catch { }
-            }
-            // All string arguments must be non-null for PolicyHit; ensure empty fallback.
-            var hit = new PolicyHit(
-                pid,
-                cul,
-                uid ?? string.Empty,
-                dname ?? string.Empty,
-                reg ?? string.Empty,
-                prod ?? string.Empty,
-                vtype ?? string.Empty,
-                score
-            );
-            bool phraseHit = false;
-            if (phraseMode)
-            {
-                // Phrase detection limited to enabled fields.
-                if (
-                    useName
-                    && !string.IsNullOrEmpty(dname)
-                    && dname.IndexOf(qStrict, StringComparison.OrdinalIgnoreCase) >= 0
-                )
-                    phraseHit = true;
-                if (
-                    !phraseHit
-                    && useDesc
-                    && !string.IsNullOrEmpty(explainText)
-                    && explainText.IndexOf(qStrict, StringComparison.OrdinalIgnoreCase) >= 0
-                )
-                    phraseHit = true;
-                if (
-                    !phraseHit
-                    && useId
-                    && !string.IsNullOrEmpty(uid)
-                    && uid.IndexOf(qStrict, StringComparison.OrdinalIgnoreCase) >= 0
-                )
-                    phraseHit = true;
-                if (
-                    !phraseHit
-                    && useReg
-                    && !string.IsNullOrEmpty(reg)
-                    && reg.IndexOf(qStrict, StringComparison.OrdinalIgnoreCase) >= 0
-                )
-                    phraseHit = true;
-            }
-            hits.Add((hit, phraseHit));
-        }
-        // Phase ordering: if phraseMode and we have phrase hits, show them first (score order), then fallback applying priority ordering to remainder.
-        var finalList = new List<PolicyHit>(limit);
-        if (phraseMode)
-        {
-            var phraseHits = hits.Where(h => h.Phrase)
-                .Select(h => h.Hit)
-                .OrderByDescending(h => h.Score)
-                .ThenBy(h => h.UniqueId, StringComparer.OrdinalIgnoreCase)
-                .ToList();
-            if (phraseHits.Count > 0)
-            {
-                foreach (var h in phraseHits.Take(limit))
-                    finalList.Add(h);
-                if (finalList.Count < limit)
-                {
-                    var remaining = hits.Where(h => !h.Phrase).Select(h => h.Hit).ToList();
-                    AppendPriorityOrdered(remaining, finalList, limit, qStrict);
-                }
-                // Post-filter refinement for Name-only search: ensure all tokens appear in DisplayName.
-                try
-                {
-                    bool nameOnlySelected = useName && !useDesc && !useReg && !useId;
-                    if (nameOnlySelected && !string.IsNullOrWhiteSpace(qExact))
-                    {
-                        var qNormAll = SearchText.Normalize(qExact);
-                        var tokens = qNormAll
-                            .Split(
-                                new[] { ' ' },
-                                StringSplitOptions.RemoveEmptyEntries
-                                    | StringSplitOptions.TrimEntries
-                            )
-                            .Distinct(StringComparer.Ordinal)
-                            .ToArray();
-                        if (tokens.Length > 0)
-                        {
-                            finalList = finalList
-                                .Where(h =>
-                                {
-                                    var dn = h.DisplayName ?? string.Empty;
-                                    if (string.IsNullOrWhiteSpace(dn))
-                                        return false;
-                                    var dnNorm = SearchText.Normalize(dn);
-                                    foreach (var t in tokens)
-                                        if (!dnNorm.Contains(t, StringComparison.Ordinal))
-                                            return false;
-                                    return true;
-                                })
-                                .ToList();
-                        }
-                    }
-                }
-                catch { }
-                return finalList;
-            }
-        }
-        // No phrase phase or no phrase hits -> just priority ordering of all hits.
-        AppendPriorityOrdered(hits.Select(h => h.Hit), finalList, limit, qStrict);
-        try
-        {
-            bool nameOnlySelected2 = useName && !useDesc && !useReg && !useId;
-            if (nameOnlySelected2 && !string.IsNullOrWhiteSpace(qExact))
-            {
-                var qNormAll = SearchText.Normalize(qExact);
-                var tokens = qNormAll
-                    .Split(
-                        new[] { ' ' },
-                        StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries
-                    )
-                    .Distinct(StringComparer.Ordinal)
-                    .ToArray();
-                if (tokens.Length > 0)
-                {
-                    finalList = finalList
-                        .Where(h =>
-                        {
-                            var dn = h.DisplayName ?? string.Empty;
-                            if (string.IsNullOrWhiteSpace(dn))
-                                return false;
-                            var dnNorm = SearchText.Normalize(dn);
-                            foreach (var t in tokens)
-                                if (!dnNorm.Contains(t, StringComparison.Ordinal))
-                                    return false;
-                            return true;
-                        })
-                        .ToList();
-                }
-            }
-        }
-        catch { }
-        // Extra filter for queries length <=2: drop candidates whose only hits are middle-of-word (not exact/prefix/boundary).
-        if (shortQuery)
-        {
-            var qLower = qStrict; // already strict lower
-            finalList = finalList
-                .Where(h =>
-                {
-                    string dn = h.DisplayName ?? string.Empty;
-                    string id = h.UniqueId ?? string.Empty;
-                    string reg = h.RegistryPath ?? string.Empty;
-                    // Determine if match is exact, prefix, or word-boundary.
-                    bool Accept(string s)
-                    {
-                        if (string.IsNullOrEmpty(s))
-                            return false;
-                        int idx = s.IndexOf(qLower, StringComparison.OrdinalIgnoreCase);
-                        if (idx < 0)
-                            return false;
-                        if (idx == 0)
-                            return true; // prefix
-                        char prev = s[idx - 1];
-                        return !char.IsLetterOrDigit(prev); // word boundary
-                    }
-                    // Keep only if at least one field satisfies prefix or boundary condition.
-                    if (Accept(dn) || Accept(id) || Accept(reg))
-                        return true;
-                    return false;
-                })
-                .ToList();
-        }
-
-        // Additional re-scoring: recompute a local ScoreMatch to align priority with non-cache search,
-        // applying AND mode and multi-token constraints uniformly.
-        try
-        {
-            var normQuery = SearchText.Normalize(query);
-            var tokens = andMode
-                ? normQuery.Split(
-                    new[] { ' ' },
-                    StringSplitOptions.RemoveEmptyEntries | StringSplitOptions.TrimEntries
-                )
-                : new[] { normQuery };
-            // Local replica of the non-cache ScoreMatch logic.
-            static int ScoreMatchLocal(string textLower, string qLower)
-            {
-                if (string.IsNullOrEmpty(qLower))
-                    return 0;
-                if (string.Equals(textLower, qLower, StringComparison.Ordinal))
-                    return 100;
-                if (textLower.StartsWith(qLower, StringComparison.Ordinal))
-                    return 60;
-                int idx = textLower.IndexOf(qLower, StringComparison.Ordinal);
-                if (idx > 0)
-                {
-                    char prev = textLower[idx - 1];
-                    if (!char.IsLetterOrDigit(prev))
-                        return 40;
-                    return 20;
-                }
-                return -1000;
-            }
-            string Norm(string s) => SearchText.Normalize(s);
-
-            bool useNameF = useName;
-            bool useIdF = useId;
-            bool useRegF = useReg;
-            bool useDescF = useDesc; // Description scoring disabled: PolicyHit does not yet include description body.
-
-            // AND mode: every token must match (score > -1000) in at least one enabled field.
-            // OR mode: keep hit only if at least one field reaches threshold (>=20).
-            var rescored = new List<(PolicyHit Hit, int Score)>(finalList.Count);
-            foreach (var h in finalList)
-            {
-                string nameN = useNameF ? Norm(h.DisplayName ?? string.Empty) : string.Empty;
-                string idN = useIdF ? Norm(h.UniqueId ?? string.Empty) : string.Empty;
-                string regN = useRegF ? Norm(h.RegistryPath ?? string.Empty) : string.Empty;
-                // Description scoring not implemented (would require extending PolicyHit).
-                bool tokenMiss = false;
-                int aggregate = 0;
-                foreach (var t in tokens)
-                {
-                    int best = -1000;
-                    if (useNameF)
-                        best = Math.Max(best, ScoreMatchLocal(nameN, t));
-                    if (useIdF)
-                        best = Math.Max(best, ScoreMatchLocal(idN, t));
-                    if (useRegF)
-                        best = Math.Max(best, ScoreMatchLocal(regN, t));
-                    if (useDescF)
-                    { /* placeholder for future desc scoring */
-                    }
-                    if (andMode)
-                    {
-                        if (best <= -1000)
-                        {
-                            tokenMiss = true;
-                            break;
-                        }
-                        // AND mode: accumulate only non-negative scores (negatives already filtered out).
-                        aggregate += Math.Max(0, best);
-                    }
-                    else
-                    {
-                        // OR mode: track only the best field score.
-                        if (best > aggregate)
-                            aggregate = best;
-                    }
-                }
-                if (tokenMiss)
-                    continue;
-
-                // Short queries (<=2): emphasize prefix/boundary; remove pure middle (score 20) matches.
-                if (shortQuery && aggregate == 20)
-                    continue;
-                // OR mode minimum acceptable score: 20 (parity with non-cache). Short queries require >=40.
-                if (!andMode && !shortQuery && aggregate <= -1000)
-                    continue;
-                rescored.Add((h, aggregate));
-            }
-            if (rescored.Count > 0)
-            {
-                // Order by score descending; preserve original order on ties (stable via index).
-                finalList = rescored
-                    .Select((x, idx) => (x.Hit, x.Score, idx))
-                    .OrderByDescending(e => e.Score)
-                    .ThenBy(e => e.idx)
-                    .Select(e => e.Hit)
-                    .Take(limit)
-                    .ToList();
-            }
-        }
-        catch { }
-        return finalList;
-    }
-
-    private static void AppendPriorityOrdered(
-        IEnumerable<PolicyHit> source,
-        List<PolicyHit> dest,
-        int limit,
-        string qLowerOrder
-    )
-    {
-        int Priority(PolicyHit h)
-        {
-            bool idHit =
-                !string.IsNullOrEmpty(h.UniqueId)
-                && h.UniqueId.IndexOf(qLowerOrder, StringComparison.OrdinalIgnoreCase) >= 0;
-            bool nameHit =
-                !string.IsNullOrEmpty(h.DisplayName)
-                && h.DisplayName.IndexOf(qLowerOrder, StringComparison.OrdinalIgnoreCase) >= 0;
-            bool regHit =
-                !string.IsNullOrEmpty(h.RegistryPath)
-                && h.RegistryPath.IndexOf(qLowerOrder, StringComparison.OrdinalIgnoreCase) >= 0;
-            bool regValueHit = false;
-            if (regHit)
-            {
-                try
-                {
-                    var parts = h.RegistryPath.Split('\\');
-                    if (parts.Length > 0)
-                    {
-                        var last = parts[^1];
-                        if (
-                            !string.IsNullOrEmpty(last)
-                            && last.IndexOf(qLowerOrder, StringComparison.OrdinalIgnoreCase) >= 0
-                        )
-                            regValueHit = true;
-                    }
-                }
-                catch { }
-            }
-            if (idHit || nameHit || regValueHit)
-                return 3;
-            if (regHit)
-                return 2;
-            return 1;
-        }
-        foreach (
-            var h in source
-                .Select(h => (h, pri: Priority(h)))
-                .OrderByDescending(t => t.pri)
-                .ThenByDescending(t => t.h.Score)
-                .ThenBy(t => t.h.UniqueId, StringComparer.OrdinalIgnoreCase)
-                .Select(t => t.h)
-        )
-        {
-            if (dest.Count >= limit)
-                break;
-            dest.Add(h);
-        }
-    }
+    ) => AdmxCacheSearchService.SearchAsync(this, query, cultures, fields, andMode, limit, ct);
 
     public Task<IReadOnlyList<PolicyHit>> SearchAsync(
         string query,
